I'm done....sort of.
I did the second assignment.  Not because I'm trying to hot-dog it....far from it.  But because the second assignment just made more sense to me.  Trying to guess a random number is an exercise in futility.....from  statistical point of view.  Anyhow it went pretty well but once again I didn't have time to get it finished.  I like to put some of me into these assignments and I guess that's taking me longer than others.  That's why I want to be a developer....it's a creative endeavor...and even though I spent the greater part of 2 days at it....I enjoyed the time I spent.  This is fun.  It's like a giant puzzle with many ways to figure it out.  I just wish I had the time to get it done the way I'd like to.