files added
Some frustration here.  I couldn't get my javascript to run.  I even put a message in there to alert me when the image was clicked and I never got the alert so something was definitely wrong because I was the one doing the clicking.  I also couldn't get my background image to display on the webpage.  I tried it two different ways.  I had several other issues I was able to clear out on my own and I'll never forget how I did each one.  It's better for me to work out these bugs on my own because I learn better working it out.  But it takes so very long to do so.  Towards the end of my time I started to wonder if my naming scheme was off in html....that I was possibly calling variables and functions for elements I'd not yet defined and that's something I wanted to go through from top to bottom and check element by element but I didn't have the time to do so.  I am very much enjoying learning all this great stuff but I want my code to run.